requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.0.0
asyncpg==0.29.0
python-dotenv==1.0.1